i18n: move to config/client folder
