+++ config/client/i18n.js
