+++ docs/i18n.md
@@ -14,7 +14,7 @@ import { withNamespaces } from '@/modules/core/client/utils/i18n-angular-load';
  * Use this one for every other case
  */
 // import { withNamespaces } from 'react-i18next';
-// import '@/config/lib/i18n'; // sometimes you'll also need to import i18n config
+// import '@/config/client/i18n'; // sometimes you'll also need to import i18n config
 
 export function MyComponent(props) { // export the unwrapped component if you want to test it
   return (
@@ -122,7 +122,7 @@ t('A translation string with {{amount}} words.', { amount: 6 })
 t('Translation created on {{ date, MMM D YYYY }}.', { date: new Date('2018-12-25') });
 ```
 
-You can see available formattings in `./config/lib/i18n.js`. And you can add more there.
+You can see available formattings in `./config/client/i18n.js`. And you can add more there.
 
 For the translations containing html, you can use [Trans](https://react.i18next.com/components/trans-component) component.
