+++ modules/references/client/components/create-reference/Recommend.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { ToggleButton, ToggleButtonGroup } from 'react-bootstrap';
-import '@/config/lib/i18n';
+import '@/config/client/i18n';
 import { withNamespaces } from 'react-i18next';
 import Report from './Report';
