+++ modules/users/client/components/ProfileViewBasics.component.js
@@ -2,7 +2,7 @@ import React from 'react';
 import classnames from 'classnames';
 import { withNamespaces } from '@/modules/core/client/utils/i18n-angular-load';
 import { Trans } from 'react-i18next';
-import '@/config/lib/i18n';
+import '@/config/client/i18n';
 import PropTypes from 'prop-types';
 import * as languages from '@/config/languages/languages';
 import { hasConnectedAdditionalSocialAccounts, isWarmshowersId, socialAccountLink } from './utils/networks';
